ENHANCEMENT: Suggestions for visualize() improvements
There are a number of example cases in this committed script (and set of quick functions) that demo some new visualization features.

Things are generally slow with the new read.pdb() data.frame that includes HETATOMs but all does appear to work ok. I guess not having CONNECT records could be an issue for some work.