OTHER: Fixed row pre-selection in SEARCH datatable. Folder 'pca-app2' is created to relfect these changes and avoid conflict with existing logic/design before a merge decision is made.
The pre-selections, and any further [de]selections made either via the interactive datatable or the hits slider are persistent when clicking the 'align' tab.

The selections are now marked by a navyblue circle (color choice made consistent in pca-tab) rather than changing the point color to green, as that results in loss of information as to whether the selection was above (red) or below (green) the cutoff.