OTHER: Added popover help tips to pca-app search tab as test.
- Moved js-related functions to ui/ui_utils.R
- Added help/question-icon on search page.
  popovers of particular id (viz. pop_blast_input) as instantiated
through the use of popoverQuestion. The icon added via popoverQuestion
will fire popovers upon clicking.

Various shiny input methods are
wrapped in a div with necessary popover info (such as
data-toggle='pop_blast_input') which contains the popover title and
content, as well as links them to the question-icon.

The popovers are dismissed when click on anything other than the
popover(s)
or the question-icon.